pyproject.toml
[project]
name = "saev"
version = "0.1.0"
description = "Sparse autoencoders for vision transformers in PyTorch"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beartype>=0.19.0",
    "datasets>=3.0.1",
    "einops>=0.8.0",
    "jaxtyping>=0.2.34",
    "marimo>=0.9.10",
    "open-clip-torch>=2.28.0",
    "pillow>=11.0.0",
    "submitit>=1.5.2",
    "torch>=2.5.0",
    "tqdm>=4.66.5",
    "tyro>=0.8.12",
    "wandb>=0.18.5",
    "webdataset>=0.2.100",
]

[tool.ruff.lint]
ignore = ["F722"]

[tool.uv]
dev-dependencies = [
    "pdoc3>=0.11.1",
]
